refactor(backend): split ai control quota helpers
